pyproject.toml
[project]
name = "project-notes"
version = "0.1.0"
description = "Create cards based on Obsydian notes using ChatGPT models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic==2.10.6",
    "pyyaml==6.0.2",
    "types-pyyaml==6.0.12.20241230",
]

[dependency-groups]
dev = [
    "faker==36.1.0",
    "mypy==1.15.0",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "ruff==0.9.6",
]


[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
    "conftest",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "smoke: business core tests",
]
addopts = [
    "-ra -q",
    "--cov=.",
    "--cov-report",
    "term-missing:skip-covered",
    "--cov-fail-under=80"
]

[tool.coverage.run]
omit = [
    "*/tests/*.py",
    "conftest/*"
]

[tool.mypy]
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
